feat: Delegate project build to the toolchain
This commit replaces the inlining of the toolchain invocation by a dead simple `npm run build`.

This effectively means that the contract that script authors need to comply to to be able to push
scripts to our platform is simplified, the compile dependencies that need to be met are:

- Define a `build` npm script
- The build script _must_ use @shopify/scripts-toolchain-as
- The build script _must_ output the generated wasm in the root of the directory
  following the format: <script_name>.wasm

Prior to this commit the CLI made heavy assumptions about the file
structure of a given project, making the whole process not transparent
to the script author.