Reduce language-specific knowledge from domain
The domain should not hardcode its logic to depend on the languages,
at least not in the form of method names.

This commit reduces such coupling in 2 ways:
1. Remove language-specific SDK classes (subclasses of
   `ExtensionPointSDK`). For tests where this is needed, we're
   using the `#for` method instead, which accepts the name of the
   language as a String.
2. Remove the the language methods on `Domain::ExtensionPoint`
   (`#rust`, `#assemblyscript`, `typescript`).

Also simplifies some tests that were failing because because they
had assertions unrelated to the thing under test.

While not perfect, this commit sets the table for future refactors.