Fix `rails new` command for Rails 7
Spring is no longer bundled by default in Rails 7. Creating an app
with `rails new --skip-spring app_name` will create a `--skip-spring`
directory.

This commit fixes that by putting the app name first. In my testing,
`rails new` ignores the extra flag it doesn't understand, so it should
work on both Rails 6 and Rails 7.

For context, Spring was removed in this commit:
https://github.com/rails/rails/commit/21c9732fae9ce0fff7adf0da610d26eeedfc3076