Update spin FQDN mechanism
The CLI is broken when using isospin.

With the transition to isospin, we no longer have the
`SPIN_WORKSPACE`, `SPIN_NAMESPACE` or `SPIN_HOST` env var, which are
required (otherwise you get an exception when trying to login).

The documented workaround is to use INFER_SPIN=1, with which the CLI
uses the `spin` binary to get the FQDN. But that doesn't work either,
because the `spin` binary is not provided in isospin.

With this change, what I'm proposing is that we
1. use a single ENV var -- `SPIN` -- to detect whether we're on spin
2. read the FQDN from the file, as recommended by Burke