Gracefully handle unknow extension surface areas
Previously, extension surface areas that weren't properly configured
would cause the CLI to error out. Now, the name of the extension gets
updated to indicate that it is not properly configured. This will let
Shopify developers working on new extensions know that further work is
required while also not getting in the way when you create extension sof
a different type.

This issue should never arrise in production. It surface because the
extensions spin template activates beta flags for extensions that are
still under development. One of these extensions still requires work in
the CLI to be done to function properly. We now gracefully degrade if we
know in advance that we can't proceed.