Retrieve latest CLI version in the background
Currently we check for a new version as a blocking process. This isn't a
good experience for CLI users, for whom the command hangs once daily
until we check for a new version.

This approach will fork the process and check in the background, saving
the result and warning the user the next time they run a command.

Note this isn't ideal for CLI environments which might run 1 command per
CLI run and never show the warning. We might choose to mitigate by
checking inline on non-TTY terminals.